Redo statbuf -> FileId
